Add package `ChrisBarber.Finance` version 0.1.0-beta (#28)
* Adding in Basic P&L calculation

First P&L UDF for basic calculation

* Update manifest

* Updated tmdl in DAXLIBs guidance

* Fix TMDL indentation

* Update package ID for ChrisBarber.Finance

* Update DAXLIB_PackageId annotation value

* Change description in manifest.daxlib

Updated the description to reflect the new functionality of the functions.

* Change folder name to lowercase

* Update manifest metadata for finance package

Updated the `authors` field to "christophermichaelbarber" and revised the `tags` to better reflect the financial focus of the project. Changed the `repositoryUrl` to a new
path under "chrisbarber.finance" and added an `icon` field with the value "/icon.png".

* Fix function name casing for ProfitAndLoss.Basic

---------

Co-authored-by: Alberto Spelta <albertospelta@users.noreply.github.com>